introducing userService class
